Refactoring: moved utils from handlers to registration and confirmation Deletes confirmation after use
